requirements.txt
requests>=2.31.0
pandas>=2.1.0
python-docx>=1.1.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytest>=7.4.0
colorama>=0.4.6
python-dateutil>=2.8.0
